sqlite schedule performence

+++ genw.sh
@@ -1,4 +1,7 @@
-rm /ztmtoosm/lol45.db
+if [ -z "$1" ]
+then
+	rm /ztmtoosm/lol45.db
+fi
 /ztmtoosm/bin/treelite /ztmtoosm/data/warszawa.osm /ztmtoosm/lol45.db 3651336
 echo $?
 /ztmtoosm/bin/osmlite /ztmtoosm/data/warszawa.osm /ztmtoosm/lol45.db
